<body>: Wrap content on narrow windows
By changing from a set width to a set max-width, content in the body
will now begin to wrap if the window is less than the max-width instead
of continuing off the side of the page (requiring a horizontal scroll).

This is what I wanted originally, just didn't test the previous commit
fully.